+++ config.ini
@@ -13,11 +13,11 @@ app_edition = 3.3.1
 token = 
 device_id = 
 map_key = 
-device_name = Xiaomi
+device_name = 
 utc = 
 uuid = 
 sign = 
-sys_edition = Default
+sys_edition = 
 
 [Run]
 exclude_points = ["117.209175,31.774432","117.208173,31.774737","117.208326,31.774448","117.206983,31.774748"]
