+++ main.py
@@ -20,6 +20,7 @@ from tools.drift import add_drift
 from gmssl import sm4
 from Crypto.Util.Padding import pad, unpad
 import time
+from tools.Login import Login
 
 """
 加密模式：sm2非对称加密sm4密钥
@@ -55,108 +56,6 @@ def decrypt_sm4(value, SM_KEY):
     decrypt_value = crypt_sm4.crypt_ecb(b64decode(value))
     return decrypt_value
 #original code above
-
-SM4_BLOCK_SIZE = 16
-
-class Login():
-    def md5_encryption(data):
-        md5 = hashlib.md5()  # 创建一个md5对象
-        md5.update(data.encode('utf-8'))  # 使用utf-8编码数据
-        return md5.hexdigest()  # 返回加密后的十六进制字符串
-    def hex_to_bytes(hex_str):
-        return bytes.fromhex(hex_str)
-    def pkcs7_padding(data):
-        return pad(data, SM4_BLOCK_SIZE)
-    def pkcs7_unpadding(data):
-        return unpad(data, SM4_BLOCK_SIZE)
-    def sm4_encrypt(plaintext, key, iv=None, mode='ECB', padding='Pkcs7', output_format='Base64'):
-        crypt_sm4 = sm4.CryptSM4()
-        key = Login.hex_to_bytes(key)
-        # 设置加密模式
-        if mode == 'ECB':
-            crypt_sm4.set_key(key, sm4.SM4_ENCRYPT)
-        elif mode == 'CBC':
-            iv = Login.hex_to_bytes(iv) if iv else None
-            crypt_sm4.set_key(key, sm4.SM4_ENCRYPT, iv)
-        # 数据填充
-        if padding == 'Pkcs7':
-            plaintext = Login.pkcs7_padding(plaintext.encode())
-        # 加密操作
-        if mode == 'ECB':
-            ciphertext = crypt_sm4.crypt_ecb(plaintext)
-        elif mode == 'CBC':
-            ciphertext = crypt_sm4.crypt_cbc(plaintext)
-        # 输出格式转换
-        if output_format == 'Base64':
-            return base64.b64encode(ciphertext).decode()
-        elif output_format == 'Hex':
-            return ciphertext.hex()
-    def sm4_decrypt(ciphertext, key, iv=None, mode='ECB', padding='Pkcs7', input_format='Base64'):
-        crypt_sm4 = sm4.CryptSM4()
-        key = Login.hex_to_bytes(key)
-        # 设置解密模式
-        if mode == 'ECB':
-            crypt_sm4.set_key(key, sm4.SM4_DECRYPT)
-        elif mode == 'CBC':
-            iv = Login.hex_to_bytes(iv) if iv else None
-            crypt_sm4.set_key(key, sm4.SM4_DECRYPT, iv)
-        # 输入格式转换
-        if input_format == 'Base64':
-            ciphertext = base64.b64decode(ciphertext)
-        elif input_format == 'Hex':
-            ciphertext = bytes.fromhex(ciphertext)
-        # 解密操作
-        if mode == 'ECB':
-            plaintext = crypt_sm4.crypt_ecb(ciphertext)
-        elif mode == 'CBC':
-            plaintext = crypt_sm4.crypt_cbc(ciphertext)
-        # 数据去填充
-        #if padding == 'Pkcs7':
-            #plaintext = pkcs7_unpadding(plaintext)
-        return plaintext.decode()
-    def main(username, password, my_uuid, DeviceName):
-        utc = int(time.time())
-        uuid= my_uuid   
-        #md5签名结果用hex
-        encryptData='''{"password":"'''+password+'''","schoolId":"100","userName":"'''+username+'''","type":"1"}'''
-        print(encryptData)
-        #签名结果
-        sign_data='platform=android&utc={}&uuid={}&appsecret=pie0hDSfMRINRXc7s1UIXfkE'.format(utc,uuid)
-        sign=Login.md5_encryption(sign_data)
-        key='e2c9e15e84f93b81ee01bbd299a31563'
-        content=Login.sm4_encrypt(encryptData, key, mode='ECB', padding='Pkcs7', output_format='Base64')
-        content=content[:-24]
-        url = "http://210.45.246.53:8080/login/appLoginHGD"
-        headers = {
-            "token": "",
-            "isApp": "app",
-            "deviceId": uuid,
-            "deviceName": DeviceName,
-            "version": "3.4.5",
-            "platform": "android",
-            "uuid": uuid,
-            "utc": str(utc),
-            "sign": sign,
-            "Content-Type": "application/json; charset=utf-8",
-            "Accept-Encoding": "gzip",
-            "User-Agent": "okhttp/3.12.0"
-        }
-        # 请求体内容
-        data = {
-            "cipherKey": "BL+FHB2+eDL3gMtv1+2UljBFraZYQFOXkmyKrqyRAzcw1R4rsq1i8p1tEOXhZMHTlFWmR+i/mdf4DNi0hCUSoQ88JMTUSUIkgU0+mowqRlVc/n/qYGqXERFqyMqn+GANUvWU65+F6/RLhpAB3AiYSJOY/RplvXmRvQ==",
-            "content": content
-        }
-        # 发送POST请求
-        response = requests.post(url, headers=headers, json=data)
-        # 打印响应内容
-        print(response.status_code)
-        result=response.text
-        DecryptedData=json.loads(Login.sm4_decrypt(result,key, mode='ECB', padding='Pkcs7',input_format='Base64'))
-        token=DecryptedData['data']['token']
-        print(token)
-        uuid = str(random.randint(1000000000000000, 9999999999999999))
-        return token
-
 
 def getsign(utc, uuid):
     sb = (
@@ -566,30 +465,41 @@ if __name__ == '__main__':
     my_app_edition = conf.get("Yun", "app_edition") # app版本（我手机上是3.0.0）
 
     # 用户信息，包括设备信息
-    #my_token = conf.get("User", 'token') # 用户token 
-    #my_device_id = conf.get("User", "device_id") # 设备id （据说很随机，抓包搞几次试试看）
+    my_token = conf.get("User", 'token') # 用户token 
+    my_device_id = conf.get("User", "device_id") # 设备id （据说很随机，抓包搞几次试试看）
     my_key = conf.get("User", "map_key") # map_key是高德地图的开发者密钥
     my_device_name = conf.get("User", "device_name") # 手机名称
-    #my_sys_edition = conf.get("User", "sys_edition") # 安卓版本（大版本）
-    if conf.get('User', 'sys_edition') == 'Default':
-        my_sys_edition = 13
-    else:
-        my_sys_edition = conf.get('User', 'sys_edition')
-    #my_utc = conf.get("User", "utc")
+    my_sys_edition = conf.get("User", "sys_edition") # 安卓版本（大版本）
     my_utc = str(int(time.time()))
-    #my_uuid = conf.get("User", "uuid")
-    my_uuid = str(random.randint(1000000000000000, 9999999999999999))
+    my_uuid = conf.get("User", "uuid")
     my_device_id = my_uuid
     my_sign = conf.get("User", "sign")
 
-    #TEST CONTENT
-    utc = int(time.time())
-    username = conf.get("Login", "username")
-    password = conf.get("Login", "password")
-    print(username + " " + password)
-    my_token = Login.main(username, password, my_uuid, my_device_name)
-    print(my_token)
-
+    if len(conf.get('User', 'token')) == 0:
+        print("config中token为空，是否尝试使用账号密码登录？(y/n)")
+        choice = input()
+        if choice == 'y':
+            #TEST CONTENT
+            utc = int(time.time())
+            username = conf.get("Login", "username")
+            password = conf.get("Login", "password")
+            my_token = Login.main(username, password, my_uuid, my_device_name)
+            if len(my_sys_edition) == 0:
+                my_sys_edition = 13
+            else:
+                my_sys_edition = conf.get('User', 'sys_edition')
+            if len(my_uuid) == 0:
+                my_uuid = str(random.randint(1000000000000000, 9999999999999999))
+                my_device_id = my_uuid
+            else:
+                my_uuid = conf.get('User', 'uuid')
+                my_device_id = conf.get('User', 'device_Id')
+            if len(my_device_name) == 0:
+                my_device_name = 'Xiaomi'
+            print("登录成功，本次登录尝试获得的token为：" + my_token + "  本次生成的uuid为：" + my_uuid)
+        elif choice == 'n':
+            print("由于缺少token退出")
+            exit()
 
     # 跑步相关的信息
     # my_point = conf.get("Run", "point") # 当前位置，取消，改到map.json
